main file added
the script takes a photograph as input, then it ask the user to select a serie of points and posterizes the photo with the colors corresponding to these points. After that, it performs erosion and dilation to simplify the campitures and give the 'macchiaioli' look. At last, it paints lines of separation between the different campitures and flips the image so that it can be printed on paper and painted over